+++ R/functions-Spectrum2.R
@@ -84,15 +84,18 @@ removeReporters_Spectrum2 <- function(object, reporters=NULL, clean=FALSE) {
 ## It also uses the "versioned" constructor in C that adds also the class
 ## version(s) (see issue #163).
 Spectra2_mz_sorted <- function(peaksCount = NULL, rt = numeric(),
-                               acquisitionNum = NA_integer_,
-                               scanIndex = integer(), tic = 0, mz = numeric(),
+                               acquisitionNum = integer(),
+                               scanIndex = integer(), tic = numeric(),
+                               mz = numeric(),
                                intensity = numeric(), fromFile = integer(),
-                               centroided = NA, smoothed = NA,
-                               polarity = NA_integer_, msLevel = as.integer(2),
-                               merged = 1, precScanNum = NA_integer_,
-                               precursorMz = NA, precursorIntensity = NA,
-                               precursorCharge = NA_integer_,
-                               collisionEnergy = NA, nvalues = integer()) {
+                               centroided = logical(), smoothed = logical(),
+                               polarity = integer(), msLevel = 2L,
+                               merged = numeric(), precScanNum = integer(),
+                               precursorMz = numeric(),
+                               precursorIntensity = numeric(),
+                               precursorCharge = integer(),
+                               collisionEnergy = numeric(),
+                               nvalues = integer()) {
     ## Argument check; make sure all is OK before calling C.
     ## Fix issue #215: remove check to allow empty spectra
     ## if (length(mz) == 0 | length(intensity) == 0 | length(nvalues) == 0) {
@@ -104,110 +107,114 @@ Spectra2_mz_sorted <- function(peaksCount = NULL, rt = numeric(),
     if (length(mz) != length(intensity))
         stop("Lengths of 'mz' and 'intensity' do not match!")
     nvals <- length(nvalues)
+    ## Initialize empty vectors.
+    empty_int <- rep(NA_integer_, nvals)
+    empty_num <- as.numeric(empty_int)
+    empty_log <- as.logical(empty_int)
     ## Now match all of the lengths to the length of nvalues.
-    if (length(peaksCount) == 0)
+    if (!length(peaksCount))
         peaksCount <- nvalues
     ## rt
-    if (length(rt) == 0) {
-        rt <- rep(NA_integer_, nvals)
+    if (!length(rt)) {
+        rt <- empty_num
     } else {
         if (length(rt) != nvals)
             stop("Length of 'rt' has to match the length of 'nvalues'!")
     }
     ## acquisitionNum
-    if (length(acquisitionNum) == 1) {
-        acquisitionNum <- rep(acquisitionNum, nvals)
+    if (!length(acquisitionNum)) {
+        acquisitionNum <- empty_int
     } else {
         if (length(acquisitionNum) != nvals)
             stop("Length of 'acquisitionNum' has to match the length of 'nvalues'!")
     }
     ## scanIndex
-    if (length(scanIndex) == 0) {
-        scanIndex <- rep(NA_integer_, nvals)
+    if (!length(scanIndex)) {
+        scanIndex <- empty_int
     } else {
         if (length(scanIndex) != nvals)
             stop("Length of 'scanIndex' has to match the length of 'nvalues'!")
     }
     ## tic
-    if (length(tic) == 1) {
-        tic <- rep(tic, nvals)
+    if (!length(tic)) {
+        tic <- empty_num
     } else {
         if (length(tic) != nvals)
             stop("Length of 'tic' has to match the length of 'nvalues'!")
     }
     ## fromFile
-    if (length(fromFile) == 0) {
-        fromFile <- rep(NA_integer_, nvals)
+    if (!length(fromFile)) {
+        fromFile <- empty_int
     } else {
         if (length(fromFile) != nvals)
             stop("Length of 'fromFile' has to match the length of 'nvalues'!")
     }
     ## polarity
-    if (length(polarity) == 1) {
-        polarity <- rep(polarity, nvals)
+    if (!length(polarity)) {
+        polarity <- empty_int
     } else {
         if (length(polarity) != nvals)
             stop("Length of 'polarity' has to match the length of 'nvalues'!")
     }
     ## centroided
-    if (length(centroided) == 1) {
-        centroided <- rep(centroided, nvals)
+    if (!length(centroided)) {
+        centroided <- empty_log
     } else {
         if (length(centroided) != nvals)
             stop("Length of 'centroided' has to match the length of 'nvalues'!")
     }
     ## smoothed
-    if (length(smoothed) == 1) {
-        smoothed <- rep(smoothed, nvals)
+    if (!length(smoothed)) {
+        smoothed <- empty_log
     } else {
         if (length(smoothed) != nvals)
             stop("Length of 'smoothed' has to match the length of 'nvalues'!")
     }
     ## msLevel
-    if (length(msLevel) == 1) {
-        msLevel <- rep(msLevel, nvals)
+    if (!length(msLevel)) {
+        msLevel <- empty_int
     } else {
         if (length(msLevel) != nvals)
             stop("Length of 'msLevel' has to match the length of 'nvalues'!")
     }
     ## merged
-    if (length(merged) == 1) {
-        merged <- rep(merged, nvals)
+    if (!length(merged)) {
+        merged <- empty_num
     } else {
         if (length(merged) != nvals)
             stop("Length of 'merged' has to match the length of 'nvalues'!")
     }
     ## precScanNum
-    if (length(precScanNum) == 1) {
-        precScanNum <- rep(precScanNum, nvals)
+    if (!length(precScanNum)) {
+        precScanNum <- empty_int
     } else {
         if (length(precScanNum) != nvals)
             stop("Length of 'precScanNum' has to match the length of 'nvalues'!")
     }
     ## precursorMz
-    if (length(precursorMz) == 1) {
-        precursorMz <- rep(precursorMz, nvals)
+    if (!length(precursorMz)) {
+        precursorMz <- empty_num
     } else {
         if (length(precursorMz) != nvals)
             stop("Length of 'precursorMz' has to match the length of 'nvalues'!")
     }
     ## precursorIntensity
-    if (length(precursorIntensity) == 1) {
-        precursorIntensity <- rep(precursorIntensity, nvals)
+    if (!length(precursorIntensity)) {
+        precursorIntensity <- empty_num
     } else {
         if (length(precursorIntensity) != nvals)
             stop("Length of 'precursorIntensity' has to match the length of 'nvalues'!")
     }
     ## precursorCharge
-    if (length(precursorCharge) == 1) {
-        precursorCharge <- rep(precursorCharge, nvals)
+    if (!length(precursorCharge)) {
+        precursorCharge <- empty_num
     } else {
         if (length(precursorCharge) != nvals)
             stop("Length of 'precursorCharge' has to match the length of 'nvalues'!")
     }
     ## collisionEnergy
-    if (length(collisionEnergy) == 1) {
-        collisionEnergy <- rep(collisionEnergy, nvals)
+    if (!length(collisionEnergy)) {
+        collisionEnergy <- empty_num
     } else {
         if (length(collisionEnergy) != nvals)
             stop("Length of 'collisionEnergy' has to match the length of 'nvalues'!")