+++ NEWS.md
@@ -11,6 +11,8 @@
   little speed improvents for isobaric quantification (see PR
   [#280](https://github.com/lgatto/MSnbase/issues/280)) <2017-11-27
   Mon>.
+- Add support to reduce the featureData for OnDiskMSnExp objects (issue
+  [#285](https://github.com/lgatto/MSnbase/issues/285)) <2017-12-11 Mon>.
 
 ## Changes in version 2.5.1
 - Update dependencies (see issue [#271](https://github.com/lgatto/MSnbase/issues/271))